Add synchronous and asynchronous execute generic methods to FluentRequest
These methods provide the ability to send an HTTP request either synchronously or asynchronously with a specified method. It improves the versatility of FluentRequest class which now can handle a variety of HTTP methods, and process the requests based on the specified method either synchronously with a return response or asynchronously with a callback.